Containerize pgo
This commit adds the centos7, rhel7, and ubi7 dockerfiles for a pgo-client
container. This container will be used to create a pgo-client deployment that
gives access to the pgo cli utility. The deployment is installed using the
ansible installer and adds relevant variables to the ansible defaults and
inventory files. The container is only installed if the
pgo_client_container_install setting in the inventory file is set to true.

The pgo-client image pulls the enviroment variables needed to connect to the
operator from existing secrets. These secrets are created durring the ansible
install (currently pgo.tls and pgouser-<username>).

The cleanup ansible tasks are updated to delete the pgo-client deployment during
an update or uninstall.

Makefile is updated to build the pgo-client image and pgo binary based on new
makefile conventions. Operator code includes shell scripts to push/pull images
from google cloud registries and dockerhub. This commit updates the scripts to
include the new container image.

Adds documentation for installation pgo-client and new inventory file variables.