Standardize time.Time formatting to use RFC3339
RFC3339 (aka ISO8601) is an easily understood, parseable timestamp
format that can be transferred between applications. Some places in
the Operator were using custom timestamps in and around where it
would be preferable to use RFC3339.

This patch updates the Operator to use RFC3339 throughout. One
exception is a function that returns a a date in the format
"YYYY-MM-DD" in relation to when a managed user password expires.