Adds default pod ant-affinity rules to all deployment specifications utilized to deploy the various pods that comprise a full PG cluster. This includes the PG primary, each PG replica, the pgBackRest dedicated repository host, as well as pgBouncer (if enabled).
The default pod anti-affinity rules
specifically attempt to schedule each
pod created for the various
deployments discussed above to a
unique pod within the Kubernetes
cluster.  By default, the anti-affinity
rules are applied using the
preferredDuringSchedulingIgnoredDuringExecution
affinity type, which means Kubernetes
will attempt to adhere to the
anti-affinity rules, but will not
guarantee that they will be enforced.  This
configuration therefore favors ensuring various
pods comprising the PG cluster can
always be scheduled, even if not always on
the desired node.

It is also possible to use the
requiredDuringSchedulingIgnoredDuringExecution
affinity type when applying default
pod anti-affinity.  The specific affinity type
for a PG cluster can be specified by setting the
'pod-anti-affinity' option available on the
'pgo create cluster' command to either 'required'
(for
requiredDuringSchedulingIgnoredDuringExecution)
or 'preferred' (for
preferredDuringSchedulingIgnoredDuringExecution).
The same option can also be set globally
for all PG deployments by using the
'PodAntiAffinity' setting in the 'pgo.yaml'
configuration file.

When the 'required' option is utilized, a unique
node will be required for each pod comprising
the full PG cluster.  Therefore, as new deployments
(e.g. pgBouncer and/or additional replicas) are
added, there must be enough nodes available without
existing PG pods scheduled and deployed for the
cluster, otherwise those deployments will not be
successfully scheduled.  Therefore, while 'required'
supports a highly-available configuration by
ensuring node failures only potentially impact
a single deployment in the cluster, it is
necessary that the cluster has sufficient nodes
to support this configuration.

And lastly, pod anti-affinity can be disabled
all together for all deployments by setting
either the 'pod-anti-affinity' CLI option to
'disable', or by setting the 'PodAntiAffinity'
setting in pgo.yaml to disabled.