restores explicit go binary build rules
The introduction of a second, similarly named implicit rule left
the "warning suppression" rule as the more explicit rule.
This had the negative effect of suppressing the binary builds.

This change reverts the removal of the explicit build rules,
leaving the implicit rule to only cover invocations of build-*
for unneeded binaries