Update pgo show pvc and remove pgo-lspvc container
The functionality for `pgo show pvc` was unstable due to the searches
on the PVC filesystem. Reports from the field indicated there was
inconsistent behavior as well (in part due to environments that
restrict multiple mountings of the same PVC).

This modifies the functionality of `pgo show pvc` to only list the PVCs
that are associated with clusters.

Additionally, this removes the `pgo-lspvc` container as it is no longer
required by `pgo show pvc`

Issue: [ch6628]