extends client trust support to include system CAs
Prior to this change, the client already had a configuration setting for
allowing more than the server's certificate to be provided as CA trust.
Though the variable (PGO_CA_CERT) is not clearly named, it allows
processing of an entire trust store using that variable.

This change loads the OS-provided CA trust store by default and
implements a new switch for disabling this default behavior.

Other cleanups as part of this change:
 - Removed unnecessary package-scoped variables and bundled like consts
 - Renamed GetTLSCredentials to reflect its actual return information
 - Removed extraneous/prediagnostic full file reads
 - Created client "new" functions for creating TLS/non-TLS http clients