Fix for cluster names with "-" in scaledown/failover query
Cluster names that had dashes (e.g. "broken-cluster") would not properly
display replica names in the `pgo scaledown --query` and
`pgo failover --query` commands. This is due to us trying to leverage
the information in Patroni to calculate values that are made available
in the output of these commands (e.g. replication lag) so we don't
reinvent the wheel, but have to contend with Patroni get the names
from the pods themselves.

The fix uses the information we've collected in the instance-node map
to look for the instance name prefix, which will be less brittle than
the regular expression I had concocted in 4fbcd2239.

Issue: [ch6673]