This change allows the configuration settings required by pgBackRest to utilize AWS S3 storage to be provided on a cluster-by-cluster basis using CLI arguments and/or associated REST endpoints, as well as via a pgTask.  The parameters that can now be provided per cluster include: 1) the AWS S3 storage bucket; 2) the AWS endpoint; 3) the AWS region; 4) the AWS key; 5) and the AWS key secret.  This allows each individual cluster created via the PostgreSQL Operator to have its own specific AWS S3 credentials.  However, please note that global S3 configuration settings can still be specified in the pgo.yaml and aws-s3-credentials.yaml files in lieu of providing them on a cluster-by-cluster basis.
When the AWS key and key secret are provided
via the CLI and/or associated REST endpoint, they
are then turned into a Kubernetes secret, which
is utilized to populate the proper pgBackRest S3
key and key secret environment variables. If a
pgtask is submitted directly without first going
through the CLI and/or API server, this secret must
first be generated in order to avoid storing these
credentials directly in the pgTask.

In addition to the above, this PR enables the
cloning of clusters that utilize AWS S3 storage
only, or a combination of both AWS S3 and local
storage.  When cloning a cluster that utilizes S3
exclusively for pgBackRest storage, pgBackRest
data is synced to a new directory (based on the
new/cloned clusters name) within the S3 storage
bucket utilized by cluster being cloned. This is
done using the AWS CLI, which is now included in
the pgo-backrest-repo-sync container.

When the cluster being cloned utilizes both local
and S3 storage, it is possible to select either
as the  data source for the pgBackRest repo sync
and restore that will occur as part of the cloning
process.  This is done using the
'pgbackrest-storage-source' option when running
the 'pgo clone' command.  When local and S3 are
enabled,  the AWS CLI will sync files between the S3
bucket and local storage, or from local
storage to the S3 target (while still using rsync,
to sync between local volumes) depending on the
data source specified.