Remove # backups from `pgo status`
The way that the statistics on the individual cluster are gathered (i.e.
`pgo show backup`), the Operator performs a `kubectl exec` into the
primary instance of a  PostgreSQL cluster and executes the
`pgbackrest info`. This info is then forwarded along to a user.

The good news: there is a `pgbackrest info --output=json` that can be
used to get the number of backups in a parseable format that can be used
to get this total.

The bad news: we would have to exec into every pod in order to provide
this statistics.

The solution: We should remove this statistic until there is a better
way for us to track total backups.

Issue: [ch6702]