Add support for the "*" permission for a pgorole
The RBAC permission list for a pgorole became quite large, and with new
endpoints and, in turn, new permissions being added, this was becoming
unwieldly. Coupled with the new installation and distribution methods,
this became fraught with peril that we may accidentally block out a
feature from being used.

This adds the new "*" permissions that allows a pgorole to access any
API server resource. This also sets it to be the default rule in all
of the installation methods.

Issue: [ch6742]