With this change, the pgBackRest dedicated repository host will no longer be restarted when the primary DB changes, e.g. when a replica is promoted during a failover event.  Instead of relying on an environment variable configured within the backrest repo pod that defines the location of the PGDATA dir on the primary, the value of the PGDATA directory is dynamically injected into the various pgBackRest commands run for stanza-create and and backup jobs.
This strategy ensures that the pgBackRest repo will
be available as needed to run certain pgBackRest
command, i.e. there is no longer the risk that a
configuration change will result in the backrest
repo restarting, leading to an error when certain
backrest jobs are executed (e.g. when a backup is
taken post-failover).