pgo-client deployment file missing fields
The deployment file was updated to `apps/v1` instead of `extensions/v1beta` in a
previous commit. This change required a new field in the deployment json file
under `spec.selector`. The field has been added to the ansible deployment
template to user `spec.selector.matchLabels` which is set to the
`spec.template.labels`.

Issue: #1171