Postgres Operator 4.2.0 Upgrade Documenation
Along with the great new features added in Operator 4.2.0,
we want to make sure we have documented steps to make it as
easy and streamlined as possible for users to upgrade to this
new version.

This commit updates and adds documenation as needed so that users
currently using Operator versions 3.5 and all subversions of 4.0 and
4.1 are able to upgrade to 4.2.0. This includes procedures to upgrade
to a 4.2.0 Bash installation for all versions, as well as upgrade
procedures for those already using an Ansible installation for their
Postgres Operator 4.X deployment.

Also, links on the main upgrade page are updated accordingly.