Execute policies directly on the PostgreSQL primary Pod
Polices that were being applied were executing a separate script that
would connect to a PostgreSQL primary over an unencrypted connection
as a PostgreSQL superuser to execute the policy. The issue is further
exacerbated if one choose to enable TLS connections over all remote
connections to the database.

To ameliorate several problems here, any SQL policy is now applied by
directly connection to the Pod and executing the command. This
eliminates both the insecure connection and allows for users to operate
correctly within a full TLS environment.

Issue: #1180