Updates the custom-config and custom-config-ssl examples as needed for the crunchy-postgres-ha container. In both examples, a custom Patroni configuration file is now created, which is utilized to configure the cluster.  For instance, it is utilized to configure the pg_hba.conf file per Patroni guidelines. As a result, the pg_hba.conf file has been removed from both  examples.  Similarly, the pg_ident.conf file is now also configured using the custom Patroni configuration file, resulting in the removal of pg_ident.conf from the custom-config-ssl example (where it is utilized). And finally, both examples now also demonstrate how to add various other custom PG configuration settings to the DCS using this configuration file.
Specific to the custom-config example, a custom setup.sql
file is utilized to create an additional user ('test2'),
as well a table in the database defined using
PGHA_DATABASE.  Additionally, this example demonstrates
the use of an existing postgresql.conf configuration file
as the base configuration file for the node (i.e. the
postgresql.conf file  specified using the custom_conf
parameter).

Specifically for the  custom-config-ssl example,
configuration has now been applied to create a client
certificate for the replication user. This certificate
is then added to the custom config configMap, where it
is then detected when configuring/bootstrapping the
'crunchy-postgres-ha' container, and then utilized in
lieu of a password for all replica authentication into
the DB as needed for streaming replication.

And lastly, the operator itself has been updated to obtain
the proper configMap, i.e. either the global or user
specified custom configuration configMap, when retrieving
custom configMap contents in order to create the default
PGHA configMap.