Automate OpenShift OLM integration
A handful of new Make targets generate, install, and verify bundles for
Kubernetes and OpenShift. All but one `operator-sdk scorecard` test is
passing.

Fully qualified image names are passed as environment variables to be
compatible with the Red Hat Marketplace and align with the pending
"Related Images" proposal for OLM.

Issue: [ch6367]
Co-authored-by: Jonathan S. Katz <jonathan.katz@crunchydata.com>