Added directions for GCP Marketplace
Added directions to guide a user to run locally with with a PostgreSQL Operator
deployed via the GCP Marketplace installer.

Separated the documentation into two sections: one for deploying the PostgreSQL
Operator via the GCP Marketplace, and the other using the Ansible installer.

Expanded on how to get keys from the Operator running via the GCP Marketplace,
how to set those keys and get the pgo binary, how to setting up access to the
API server via a user, and adding namespaces to create a new PostgreSQL with the
Operator.