Ensure proper PostgreSQL instance name appears after failover
After a failover occurs, the `--query` method found on `pgo scaledown`
and `pgo failover` could potentially show the same name for every
single replica, due to it utilizing just the prefix. This of course, is
not good when one may need to failover or scale down in the future.

This fix moves the check over to a regular expression that follows the
Kubernetes pod name.

Issue: [ch6908]