Fixed race condition with adding schedules
The multi-namespace functionality created a situation in the scheduler
where an informer would be created twice on the same namespace. This
would result in two entries in the schedule table being added, which
then resulted in only one of the schedule entries being removed. As such
"phantom scheduled events" would occur.

This patch borrows a pattern from the controllers that creates a mutex
on the informer creation function and only allows a single informer to
be created on a namespace at a given time.

Issue: [ch6937]
Issue: #1215