Updates add-targeted-namespace script
- Removes dependency on expenv by using sed to replace environment variables
- If a namespace existed in the environment we need to confirm that it has
the correct labels. This is done by blindly adding the labels to the existing
namespace. These commands are moved inside of the if statement so that they
are only called if the namespace exists. If the namespace doesn't exist it
is created with the labels so adding them again adds unnecessary errors.
Moving the lines also makes the script and the purpose of these commands
more clear.