Updates backrest ssh key to use ed25519
- Removes ssh key generation from bash and ansible install
- Updates ssh key generation to ed25519
- Updates backrest scripts for ed25519
- Updates operator pgo-backrest config files

Keys are currently generated in the operator code and in both the bash and
ansible installers. The ansible installer gerenates not only the id_rsa key but
id_ecdsa and id_25519 keys as well. The bash installer only generates the id_rsa
key. The backrest-repo-config secret is created using the ssh key gereration in
the golang code instead of the keys generated by the installers. This commit
removes the key generation from the installers.

The backrest-repo-config secret is updated to used an ed25519 ssh key instead of
rsa. The fucntion to gereate a public/private ssh keypair is updated to generate
the new key. The new keys are used in place of the original rsa key in the
relevant places. The files that require the ssh key are updated to use the new
file names based on ed25519.  This includes the sshd_config and ssh config.
enter the commit message for your changes. Lines starting # with '#' will be
ignored, and an empty message aborts the commit.