Add an `--enable-autofail` flag to `pgo update cluster`
While `--disable-autofail=false` would work in enabling autofail,
it was generally confusing to users. Having the `--enable-autofail`
flag makes it clear that one can return a PostreSQL cluster to using
HA.

This also introduces an API change to the update cluster ressource,
as the "Autofail" request parameter is turned into an integer. This
avoids having to deal with a "3-value boolean", which could get
messy.

Issue: [ch7023]
Issue: #1236