Update NFS PV creation script to mimic "production" environments
The legacy NFS PV (persistent volume) creation script would expect
all of the PVs to exist in the same directory on the NFS mount point.
While this provides certain conveniences when developing and debugging
the Operator, it does not necessarily mimic production environments, and
could lead to creating some weird situations when trying to plan out
new features. Additionally, most PV provisioners do not do this as well,
for convenience and security as well.

This updates the "create-pv-nfs.sh" script to create a directory for
each PV that is created on the NFS filesystem. The script is very
opinionated (and could use a few more guardrails) and is still not
recommend for production use.

The legacy version of the script has been renamed to
"create-pv-nfs-legacy.sh" and is still available should one want to use
it. Likely, this will be removed in the future as we remove reliance on
"expenv"