Ensure all failover/scaledown targets are listed
The previous commit to address this issue (323c3749) solved most of the
cases, but pod names with shorter random characters at the end would
not be included in the `pgo scaledown --query` / `pgo failover --query`
lists.

This patch uses the character classes that are employed by the Kuernetes
random string generation (i.e. the SafeEncodeString function[1]) to
detect the name of the Pod to appropriately display the PostgreSQL
instance name in a cluster.

Issue: #1247
Issue: [ch6955]
Issue: [ch7072]

[1] https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apimachinery/pkg/util/rand/rand.go