Respect the PasswordLength value in the configuration
The value of PasswordLength in the "pgo.yaml" file was not being
honored due to a variety of factors in the code. Additionally, a
lot of different values were being chosen as default password lengths
throughout the code base.

This fixes the PasswordLength attribute from not being honored. If
the PasswordLength is not set, a defalut value of "24" is used.

Additionally, the default of 24 is used if the PasswordLength value
is invalid, and the server is the final arbiter of the password length.

Issue: [ch3212]