Add support for PostgreSQL tablespaces
Tablespaces can be used to spread out PostgreSQL workloads across
multiple volumes, which can be used for a variety of use cases:

- Partitioning larger data sets
- Putting data onto archival systems
- Utilizing hardware (or a storage class) for a particular database
object, e.g. an index

and more.

Tablespaces can be created via the `pgo create cluster` command using
the `--tablespaces` flag. The arguments to `--tablespaces` can be passed
in via a command-separated list; each argument consister of two values:

- The name of the tablespace
- The type of storage to use for the tablespace

For example:

  pgo create cluster hacluster --tablespaces=ts=nfsstorage

All tablespaces are mounted in the `/tablespaces` directory. The
PostgreSQL Operator manages the mount points and persistent volume
claims (PVCs) for the tablespaces, and ensures they are available
throughout all of the PostgreSQL lifecycle operations, including:

- Provisioning
- Backup & Restore
- High-Availability, Failover, Healing
- Clone

etc.

Two additional values were added to the pgcluster CRD:

- TablespaceMounts: a map of the name of the tablespace and its
associated storage.

Tablespaces are automatically created in the PostgreSQL cluster. You can
access them as soon as the cluster is initialized. For example, using
the tablespace created above, you could create a table on the tablespace
`ts` with the following SQL:

  CREATE TABLE (id int) TABLESPACE ts;

Presently, tablespaces can only be added to a PostgreSQL cluster when
it is initialized. Based on usage, future work will look to making this
more flexible. Dropping tablespaces can be tricky as no objects must
exist on a tablespace in order for PostgreSQL to drop it (i.e. there is
no DROP TABLESPACE .. CASCADE command).

Co-authored-by: Brian Faherty <anothergenericuser@gmail.com>
Issue: [ch6495]