Rewrite the pgBouncer implementation
This commit helps to create a single-source-of-truth for how a pgBouncer
deployment is managed alongside a PostgreSQL cluster. The logic for
managing pgBouncer is now consolidated in the Operator itself, primarily
in the "operator/cluster/pgbouncer.go" file.

Some of these changes include:

- The infrastructure scripts for pgBouncer were moved into the
PostgreSQL HA container itself
- There is a clear delineation and hierarchy of functions and how they
are called
- Plaintext passwords are no longer stored on CRDs or other insecure
objects

etc.

This has the added benefit of bringing this one step closer to a
TLS-only workflow for PostgreSQL clusters that are managed by the
PostgreSQL Operator.

This depends on CrunchyData/crunchy-containers#1190