Only Delete Backup pgtask if One Already Exists
When cleaning the existing backup pgtask Custom Resource prior to
creating a new backup for a cluster, the PostgreSQL Operator will now
only attempt to delete a prior backup pgtask if one currently exists.
This ensures that certain backups can be properly submitted and
executed in the event that the pgtask from the previous cluster backup
has been removed (either manually or by the Operator).  For instance,
this ensures that the post-restore backup executed following a cluster
restore can still be executed in the event that the user has scaled
down the cluster prior to the restore, since successfully scaling down
a replica currently results in the deletion of all pgtasks for the
specific cluster.