Remove unused installation variables
The changes to the HA system introduced in Operator 4.2 removed
the need for some global configuration settings. However, this
could cause some confusion as these artifacts are still present in
some of the configuration files and installers.

This removes those artifacts, and also introduces the
`disable_auto_failover` parameter to the Ansible installation method.