Add "pgo show pgbouncer" and associated API endpoint
A lot of the helpful information around pgBouncer deployments is hidden
or strewn about the Operator. This introduces an endpoint and command,
"pgo show pgbouncer" that surfaces relevant information around the
deployment of a pgBouncer instance amongst a PosgreSQL cluster,
including:

- The PostgreSQL cluster name
- The pgBouncer Service name
- Any IP addresses associated with the pgBouncer Service
- The credentials of the pgBouncer service account

This is part of a campaign to eliminate the need to supply a password
to pgBouncer, as the "pgbouncer" user is a service account and does
not need to be accessed by regular usage, outside of interfacing with
pgBouncer directly (e.g. for stats, modifying settings, etc.)

This also includes a few refactors to increase the reuse of code,
including an improvement to how clusters for the other pgBouncer
commands (pgo create pgbouncer, pgo delete pgbouncer) are queried.