Update to Readiness & Liveness Probe Paths
Updated the paths defined in the cluster-deployment.json template for
executing readiness and liveness probe scripts to account for the
'health' subdirectory now included under the '/opt/cpm/bin' directory
of the crunchy-postgres-ha container.  This aligns the cluster
deployment template with the reorganization of the various scripts
included in the crunchy-postgres-ha container.