Updates to the PGHA ConfigMap
As a result of recent changes to the crunchy-postgres-ha container,
all applicable PostgreSQL default settings are now set within the
container itself.  As a result, the 'postgres-ha.yaml' file previously
created by the Operator and stored within a configMap for each PG
cluster in order to configure certain PG defaults is no longer needed.
Therefore, this commit removes any logic within the  Operator for
creating and managing the default 'postgres-ha.yaml' configuration file.

To align with this change, the name of the PGHA configMap itself has
also been changed from <cluster-name>-pgha-default-config to
<cluster-name>-pgha-config to indicate that it is simply a configMap
for the PGHA cluster, and is not responsible for setting default PG
configuration settings.  Additionally, the the various functions,
logic, etc. utilized to create and manage this configMap, have also
been updated to no longer reflect the setting of any "default"
configuration settings.