Add `pgo update pgbouncer --rotate-password`
As a previous commit removed the ability for a plaintext pgBouncer
password to be passed around, it did become imperative to have a
replacement method for being able to handle a task such as updating
a pgBouncer password. As such, having the PosgreSQL Operator coordinate
the rotation of a pgBouncer password seemed to be a way to ensure the
security of the password while allowing for a user to change it.

When `pgo update pgbouncer --rotate-password` is called, it creates a
task CR entry that allows for the Operator to generate and update the
password in both the PostgreSQL deployment as well as the pgBouncer
deployment. It's necessary to restart the pgBouncer pods after the
update (read: delete), which is handled by detecting when the Secret
updates have reached the pgBouncer Pod.