Rewrite "pgo create user" and begin initial refactor of user management
As part of the effort to allow the Operator to manage PostgreSQL clusters that
run in TLS only mode, the user management functions require some changes to
ensure that any connections to the database are made using a secure connection,
in particular, exec'ing directly to the Pod and using peer authentication.

The first step in updating the user management functions is modifying the
"pgo create user" workflow to ensure it can work in a TLS-only environment.
There are several side effects of these changes as well, including:

- Defaulting the "password expiration" function to being unlimited
- Updating the create user API calls to return structured data. The information
is now output in a more user friendly manner, and returns helpful information
such as when the password expires
- Introducing the notion of PostgreSQL system accounts and ensuring a user
cannot modify one of these directly

This includes a couple of refactors building on top of the work that was done
for pgBouncer.

This also serves as the building block for the next series of patches that aim
to clean up the user management workflow.

Issue: [ch7131]