Rewrite "pgo update user"
Continuing on the effort to allow for the PostgreSQL Operator to manage TLS-only
PostgreSQL clusters, the "update show user" command was rewritten to not rely
on making database connections other than those that occur using PostgreSQL's
peer authentication from within the Pod of the primary instance.

This commit modifies the the "pgo update user" API and now returns output in a
structured format using the standard user return. Additionally, a few more
options are now available on the "pgo update user" command, including:

- `--disable-login` which disables the ability for a PostgreSQL user to login
- `--enable-login` which enables the ability for a PostgreSQL user to login
- `--expired` is now treated as an integer
- `-o` which allows for raw JSON output
- `--rotate-password`, which has the PostgreSQL Operator change the password
for the specified user accounts on the specified clusters
- `--valid-always` which sets a password to always be valid, i.e. it has no
expiration

The "pgo update user" command has two workflows: a regular update, and one where
"--expired" is set to true. One can now also use the "--expired" flag in
conjunction with "--valid-days" to reset the expiration date for a password.
If nothing is passed into `--valid-days`, the server default is used.

Issue: [ch7131]