Rewrite "pgo delete user"
Continuing on the effort to allow for the PostgreSQL Operator to manage TLS-only
PostgreSQL clusters, the "pgo delete user" command was rewritten to not rely
on making database connections other than those that occur using PostgreSQL's
peer authentication from within the Pod of the primary instance.

The trick with this one is that a PostgreSQL role can only be dropped so long
as it does not have any objects that own it. As such, it's necessary to iterate
through every connectable database and drop any object owned by this role.
This maintains the previous behavior of "pgo delete user"

(Note that this does not drop databases or tablespaces)

This commit modifies the the "pgo delete user" API and now returns output in a
structured format using the standard user return. Additionally, a few more
options are now available on the "pgo delete user" command, including:

- `-o` which allows for raw JSON output

Issue: [ch7131]