Remove Ansible 2.8 feature and bump min version to 2.7
7a2dbbcf inadvertently introduced a feature that was added to
Ansible 2.8 that allows one to quiet warnings in assertions.
While handy, this raised the compatibility to 2.8

As there is currently an installation project going on that is
looking at all these, it seems the best course of action is to
allow for people using older versions of Ansible to be able to
install the Operator, so the "quiet" statements are removed.

While we would need to do a full assessment of Ansible version
compatibility, it is safe to see at least version 2.7, given
2.7 was released in 2018, so bump the the minimum version

Issue: #1292