Enable Standby Cluster Creation with S3 Storage
This commit enables standby cluster creation, specifically for those
using AWS S3 storage.  AWS S3 provides a central pgBackRest repository
that can accessed by multiple clusters living in separate Kubernetes
clusters.  This therefore allows the PG cluster that is currently
accepting writes to push WAL to a central repo (S3), where it can then
be obtained and replayed by one or more standby clusters.  Being that
AWS S3 is utilized as part of this architecture, no direct
communication between clusters is required.  However, please note that
this does mean that streaming replication is not currently supported
for standby clusters.

In support of this feature, two new flags have been added to the
"pgo create cluster" command, which are as follows:

- standby: boolean that determines whether or not the cluster should be
created as a standby cluster
- backrest-repo-path: Allows the user to override the pgBackRest
repository path for a cluster.  While this setting can now be utilized
when creating any cluster, it is typically required for the creation
of standby clusters, since the repo path setting will need to match
that of a remote/external PG database, and will therefore likely need
to be changed from the current default.

The flags referenced above have corresponding parameters in the
in both the apiserver and operator API's, therefore allowing standby
clusters to be created whether using the pgo client, apiserver or if
submitting pgcluster CR's directly (ensuring the proper S3
configuration is provided as needed for pgBackRest S3 support).

When creating standby clusters, the lifecycle for cluster
initialization is modified.  Specifically, since the the S3 repo
already exists and is being managed by a separate PG cluster/database,
no stanza creation is needed in the S3 repo.  However, if local and
S3 storage are selected for the standby cluster, then a stanza will
be created in the local repo in an "offline" mode, as supported by
pgBackRest ("offline" mode is needed because the remote primary is not
directly accessible from the standby).

Additionally, no backups will be created during standby cluster
creation, since the primary and any cascading replicas will
bootstrapped from the shared S3 repo.  And finally, please note that
any local repo created will be empty and local to standby cluster only,
i.e. it will not contain the contents of any remote/external
pgBackRest repo (and will therefore only start to be populated with
archives and backups once standby mode is disabled for the cluster and
it is able to accept writes).

Also, please note that this commit standardizes on the use of
env var PGBACKREST_REPO1_PATH instead of PGBACKREST_REPO1_PATH to
ensure consistent configuration of pgBackRest across the Operator.