Refactor the Pod Controller
This commit refactors and restructures the Pod controller.
Specifically, the onUpdate function now implements a handler pattern
for dispatching Pod update events to the proper business logic.  Like
the Job controller, handlers have been split into various files
according to the type of event that triggered an update to the Pod.
This includes handlers for the initialization of Pods (or
re-initialization in the case of a restore) as part of the the
PG cluster creation process, a handler for the promotion of a PG
replica Pod to the primary database within the cluster, as well as a
handler for upgrades.

This cleanly separates the logic needed to route various Pod update
events while also establishing a consistent pattern for handling said
events.  Additionally this greatly reduces the complexity of current
onUpdate processing (i.e. the removal of deeply nested if statements).