Remove `--series` flag from `pgo create cluster`
This was mostly a nuisance to the API, as "Series" was a required
parameter, and one would normally set it to 1. Most users create
PostgreSQL clusters ad hoc, and if they need to create a series of
them, as the flag suggests, they can just script it using external
tools.

Issue: [ch7230]