Allow cluster PVC size to be set at cluster creation
The storage spec files that are available for configuration of cluster
PVCs provide a "size" attribute that indicates how large the PVC should
be. However, if users want to create a PostgreSQL cluster of a new size,
they are required to create a new storage spec, which is tedious.

This patch adds the ability to select the size of the PVC when creating
the PostgreSQL cluster. This introduces the "--pvc-size" flag to
"pgo create cluster" that allows one to specify a Kubernetes parseable
"quantity" and the PostgreSQL Operator will then set the PVC size for
the primary and replica storage.

Note that this treats the primary and replica storage as the same size
when "--pvc-size" is set. This will undoubtedly lead to requests for
allowing for storage resizing, which will be handled in the future.

Issue: [ch7227]
Issue: #1080