Replace tablespace flag with --tablespace and update interface
The "--tablespaces" flag is removed and replaced with "--tablespace"
so that a user can specify adding multiple tablespaces in this way:

    pgo create cluster hacluster \
        --tablespace=ts1=nfsstorage
        --tablespace=ts2=nfstorage

which conforms to more of the expected behavior amongst CLI enthusiasts.

The other motivation of this was to move the API portion of creating
a tablespace to use structured data vs. string parsing. Any string
parsing would occur on the client side (which is what the pgo client
now does) and the server can now simply manage structured data.